requirements.txt
fastapi==0.120.2
uvicorn==0.35.0
python-multipart==0.0.20
llama_index==0.14.6
moviepy==2.2.1
pydantic==2.12.3
pydantic_settings==2.11.0
qdrant_client==1.15.1
minio==7.2.18
redis==5.3.1
youtube_transcript_api==1.2.3
yt_dlp==2025.10.22
llama-index-embeddings-clip==0.5.1
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-llms-ollama==0.8.0
llama-index-llms-openai==0.6.5
llama-index-multi-modal-llms-openai==0.6.2
llama-index-storage-docstore-redis==0.4.1
llama-index-storage-index-store-redis==0.5.1
llama-index-vector-stores-qdrant==0.8.6
git+https://github.com/openai/CLIP.git
openai-whisper==20250625
